+++ tests/Unit/CapiRemediationTest.php
@@ -54,7 +54,7 @@ use org\bovigo\vfs\vfsStreamDirectory;
  * @uses   \CrowdSec\RemediationEngine\Configuration\AbstractRemediation::addGeolocationNodes
  * @uses   \CrowdSec\RemediationEngine\AbstractRemediation::getCountryForIp
  * @uses \CrowdSec\RemediationEngine\Configuration\AbstractCache::addCommonNodes
- * @uses \CrowdSec\RemediationEngine\AbstractRemediation::updateRemediationOriginCount
+ * @uses \CrowdSec\RemediationEngine\AbstractRemediation::updateMetricsOriginsCount
  *
  * @covers \CrowdSec\RemediationEngine\AbstractRemediation::getCacheStorage
  *