+++ CHANGELOG.md
@@ -29,8 +29,8 @@ As far as possible, we try to adhere to [Symfony guidelines](https://symfony.com
 ### Changed
 
 - **Breaking change**: `getIpRemediation` method now returns an array with `remediation` and `origin` keys
-- **Breaking change**: Rename `AbstractRemediation::updateRemediationOriginCount` method to 
-  `updateRemediationOriginCount` with new `$remediation` and `$delta` parameters.
+- **Breaking change**: Change protected `AbstractRemediation::updateRemediationOriginCount` method to public
+  `updateMetricsOriginsCount` with new `$remediation` and `$delta` parameters.
 - **Breaking change**: Store `clean` as origin in cache for `bypass` remediation even if 
   original retrieved remediation was not a bypass (unhandled or capped remediation ). If `bypass` is the result of 
   AppSec remediation, we continue to store `clean_appsec` as origin in cache.